+++ csw/src/main/java/br/com/sarc/csw/modules/reserva/repository/ReservaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.sarc.csw.modules.reserva.model.Reserva;
 
-public interface ReservaRepository extends JpaRepository<Reserva, Integer> {
-    
+public interface ReservaRepository extends JpaRepository<Reserva, Long> {
+
 }
